+++ test/Datadog.Trace.ClrProfiler.IntegrationTests/AdoNet/NpgsqlCommandTests.cs
@@ -59,7 +59,7 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.AdoNet
 
             // NOTE: opt into the additional instrumentation of calls into netstandard.dll
             // see https://github.com/DataDog/dd-trace-dotnet/pull/753
-            SetEnvironmentVariable("DD_TRACE_NETSTANDARD_ENABLED", "true");
+            SetEnvironmentVariable("OTEL_TRACE_NETSTANDARD_ENABLED", "true");
 
             int agentPort = TcpPortProvider.GetOpenPort();