+++ test/Datadog.Trace.ClrProfiler.IntegrationTests/Helpers/TestHelper.cs
@@ -231,13 +231,13 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests
 
         protected void SetServiceVersion(string serviceVersion)
         {
-            SetEnvironmentVariable("DD_VERSION", serviceVersion);
+            SetEnvironmentVariable("OTEL_VERSION", serviceVersion);
         }
 
         protected void SetCallTargetSettings(bool enableCallTarget, bool enableMethodInlining)
         {
-            SetEnvironmentVariable("DD_TRACE_CALLTARGET_ENABLED", enableCallTarget ? "true" : "false");
-            SetEnvironmentVariable("DD_CLR_ENABLE_INLINING", enableMethodInlining ? "true" : "false");
+            SetEnvironmentVariable("OTEL_TRACE_CALLTARGET_ENABLED", enableCallTarget ? "true" : "false");
+            SetEnvironmentVariable("OTEL_CLR_ENABLE_INLINING", enableMethodInlining ? "true" : "false");
         }
 
         protected async Task AssertWebServerSpan(